backend/requirements.txt
# Core Flask framework
Flask==2.3.3
python-dotenv==1.0.0

# AI/ML Libraries
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.25.2
tensorflow==2.13.0
torch==2.0.1
transformers==4.34.0

# Explainable AI
shap==0.42.1
lime==0.2.0.1

# Natural Language Processing
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1
gensim==4.3.2

# Data Processing & Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1
wordcloud==1.9.2

# File Processing
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2

# Web scraping (for job data collection)
requests==2.31.0
beautifulsoup4==4.12.2

# Database (optional)
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.1.1

# Development & Testing
pytest==7.4.2
pytest-flask==1.2.0
jupyter==1.0.0

# Performance & Caching
redis==5.0.0
celery==5.3.2

# Additional utilities
joblib==1.3.2
python-json-logger==2.0.7
tqdm==4.66.1
